--- a/package.json
+++ b/package.json
@@ -29,8 +29,8 @@
   "devDependencies": {
     "@babel/core": "^7.14.3",
     "@babel/preset-env": "^7.14.5",
-    "@testing-library/react": "^12.0.0",
-    "@types/jest": "^26.0.23",
+    "@testing-library/react": "^12.1.3",
+    "@types/jest": "^27.4.1",
     "@types/plotly.js": "^1.54.20",
     "@types/react": "^17.0.9",
     "@types/react-dom": "^17.0.6",
@@ -39,13 +39,13 @@
     "@typescript-eslint/parser": "^4.26.1",
     "esbuild-loader": "^2.18.0",
     "eslint": "^7.28.0",
-    "jest": "^27.0.4",
+    "jest": "^27.4.1",
     "jest-canvas-mock": "^2.3.1",
-    "prettier": "^2.3.1",
-    "ts-jest": "^27.0.3",
-    "ts-loader": "^8.3.0",
-    "typescript": "^4.3.2",
-    "webpack": "^5.38.1",
-    "webpack-cli": "^4.7.2"
+    "prettier": "^2.5.1",
+    "ts-jest": "^27.1.3",
+    "ts-loader": "^9.2.7",
+    "typescript": "^4.6.2",
+    "webpack": "^5.70.0",
+    "webpack-cli": "^4.9.2"
   }
 }