--- a/webpack.config.js
+++ b/webpack.config.js
@@ -2,6 +2,7 @@ const webpack = require('webpack');
 
 const mode = process.env.NODE_ENV === 'production' ? 'production' : 'development';
 const isDev = mode === 'development';
+const skipTypeCheck = process.env.SKIP_TYPE_CHECK === 'true';
 
 var config = {
     mode,
@@ -29,7 +30,6 @@ var config = {
     },
     plugins: [
         new webpack.DefinePlugin({
-            'process.env.NODE_ENV': JSON.stringify(process.env.NODE_ENV),
             'APP_BAR_TITLE': JSON.stringify(process.env.APP_BAR_TITLE || "Optuna Dashboard"),
             'API_ENDPOINT': JSON.stringify(process.env.API_ENDPOINT),
             'URL_PREFIX': JSON.stringify(process.env.URL_PREFIX || "/dashboard")
@@ -41,6 +41,20 @@ if (isDev) {
     config.devtool = 'source-map';
     console.log('= = = = = = = = = = = = = = = = = = =');
     console.log('DEVELOPMENT BUILD');
+    console.log('= = = = = = = = = = = = = = = = = = =');
+}
+if (skipTypeCheck) {
+    config.module.rules = [{
+        test: /\.tsx?$/,
+        exclude: [/node_modules/],
+        loader: 'esbuild-loader',
+        options: {
+            loader: 'tsx',
+            tsconfigRaw: require('./tsconfig.json')
+        }
+    }]
+    console.log('= = = = = = = = = = = = = = = = = = =');
+    console.log('Use esbuild-loader instead of ts-loader');
     console.log('= = = = = = = = = = = = = = = = = = =');
 }