<!--
Thank you for creating a pull request!
-->

## Contributor License Agreement

This repository (``optuna-dashboard``) and [Goptuna](https://github.com/c-bata/goptuna) share common code.
This pull request may therefore be ported to Goptuna.
Make sure that you understand the consequences concerning licenses and check the box below if you accept the term before creating this pull request.

* [x] I agree this patch may be ported to [Goptuna](https://github.com/c-bata/goptuna) by other Goptuna contributors.

## Reference Issues/PRs
<!--
Example: Fixes #1234. Close #1234. See also #1234. Follow-up #1234.
Please use keywords (e.g., Fixes) to automatically close referenced issues.
-->
None

## What does this implement/fix? Explain your changes.

<!-- Describe the changes in this PR. A picture or video tells a thousand words. -->

I referred https://webpack.js.org/guides/build-performance/ to improve build performance.
 
* [x] Update deps
* [x] Enable build cache
* [x] Use esbuild-loader by default

I compared the time consumed for TypeScript builds.

| Option | Build time (`npm run build:dev`) |
| --- | --- |
| ts-loader with `transpileOnly=true` | 25955ms | 
| ts-loader with `transpileOnly=true HappyPackMode=true` | 29804ms |
| esbuild-loader | 20183ms |
| ts-loader with `transpileOnly=true` and `fork-ts-checker-webpack-plugin` | 30185ms |